requirements.txt
Django==1.10
redis-py==2.12.4
sorl-thumbnail==12.3